refactor(ember): extract internal tabs primitive, split docfy-demo
Extracts the registration/selection logic behind docfy-demo's snippet
tab strip into an internal Tabs primitive at src/-private/tabs.gts,
deliberately outside components/ so it is not re-exported into
consuming apps and not part of the public package surface. Splits the
250-line docfy-demo.gts into a components/docfy-demo/ folder
(index, description, example, snippet, snippets), with snippets.gts
now built on top of Tabs while still rendering its own legacy markup
and data-test-ids so existing tests keep passing unchanged.

Pure refactor, no observable behaviour change.
