+++ packages/ember/package.json
@@ -76,7 +76,11 @@
     "type": "addon",
     "main": "addon-main.cjs",
     "app-js": {
-      "./components/docfy-demo.js": "./dist/_app_/components/docfy-demo.js",
+      "./components/docfy-demo/description.js": "./dist/_app_/components/docfy-demo/description.js",
+      "./components/docfy-demo/example.js": "./dist/_app_/components/docfy-demo/example.js",
+      "./components/docfy-demo/index.js": "./dist/_app_/components/docfy-demo/index.js",
+      "./components/docfy-demo/snippet.js": "./dist/_app_/components/docfy-demo/snippet.js",
+      "./components/docfy-demo/snippets.js": "./dist/_app_/components/docfy-demo/snippets.js",
       "./components/docfy-link.js": "./dist/_app_/components/docfy-link.js",
       "./components/docfy-output.js": "./dist/_app_/components/docfy-output.js",
       "./components/docfy-previous-and-next-page.js": "./dist/_app_/components/docfy-previous-and-next-page.js",
