+++ packages/ember/src/index.ts
@@ -1,4 +1,4 @@
-export { default as DocfyDemo } from './components/docfy-demo.gts';
+export { default as DocfyDemo } from './components/docfy-demo/index.gts';
 export { default as DocfyLink } from './components/docfy-link.gts';
 export { default as DocfyOutput } from './components/docfy-output.gts';
 export { default as DocfyPreviousAndNextPage } from './components/docfy-previous-and-next-page.gts';
